Improve multi-error handling and move the logic to parsec.utils
